Bugfix: Storage auf 32 bit festgelegt

+++ include/Storage.h
@@ -1,11 +1,13 @@
 #ifndef STORAGE_H_
 #define STORAGE_H_
 
+#include <QtGlobal>
+
 //! Datenstruktur der Speicherzelle
 union Storage
 {
-    int i;
-    double f;
+    qint32 i; // garantiert 32 bit auf jeder von Qt unterstützten Plattform
+    float f;  // per IEEE-754 Definition 32 bit auf jeder Plattform.
 };
 
 #endif /*STORAGE_H_*/
